3DS: Implement graphics modes corresponding to hardware pixel formats
When launching a game, do the following:
1) Check if the graphics mode currently in use is most optimal for that game.
2) If not, do the following:
2a) Switch to the most appropriate graphics mode.
2b) Unload render targets and shaders with unloadGfx().
2c) Set the LCD screens to the appropriate pixel format with gfxSetScreenFormat().
2d) Reinitialize textures, render targets and shaders under the new graphics mode.

To Do:
- Fix virtual keyboard visibility: when opening virtual keyboard under an
alpha-less graphics mode, the bottom screen (or top screen, if ScummVM is
set to only use the top screen) will turn black except for the blue text
on the text entry line. Pressing Y or DPAD-RIGHT will close the keyboard
and restore normal screen visibility.
- Simulate transparency when the overlay is visible.